chore(conformance): delete the dead v1 toolcalling table module
With the v1 page gone, tables/toolcalling/table.py was unreachable except for
five version helpers: the conformance generator referenced only _impl_versions,
_pinned_versions, _version_slug, _version_sort_key, _fixtures_cache_root and two
path constants. The other 38 top-level defs were the v1 page's table builder.

Those 38 were also near-duplicates of the v2 copies already in fixtures.py
(load_all_cases, BATCH_SUB_CASE_GROUPS, SPLIT_PARENT_SUBCASES, _derive_no_peer_sets,
the sub-case sort/band helpers). conformance/README.md told authors to "edit BOTH"
copies when adding a case; deleting the dead one removes that hazard, so the README
TODO goes with it.

The five survivors move to fixtures.py, which already owns fixture loading and the
sub-case taxonomy. reasoning/table.py loses its two dead functions (_make_jinja_env,
_commit_sha) and the now-unused jinja2 + subprocess imports.

Net -940 lines. The rendered page is unchanged: CONFORMANCE_v2.html is byte-identical
to the pre-change render, check.sh ci passes, 88 pytest tests pass.
