+++ conformance/utils/src/fixtures.py
@@ -10,6 +10,7 @@ rendering code.
 """
 import copy
 import json
+import os
 import re
 from pathlib import Path
 
@@ -26,6 +27,68 @@ from markers import (
 REPO_ROOT = Path(__file__).resolve().parents[2]
 FIXTURES = REPO_ROOT / "tests/parity/toolcalling/fixtures"
 RUST_TOOL_CALLING_DIR = REPO_ROOT / "lib/parsers/src/tool_calling"
+
+# The versioned fixture source (inputs/ + per-impl <impl>-<version>/ dirs) lives in
+# the fixture extraction cache (from the in-repo LFS store). `_common.sh` exports
+# CONFORMANCE_FIXTURES_ROOT (the cache root); fall back to the standard cache path for
+# standalone runs. Powers the per-impl version radios: the generator resolves each
+# version snapshot and re-runs the load path so cell keys align exactly with the
+# rendered (pinned) table.
+_FRONTEND_CRATES_ROOT = Path(os.environ.get("FRONTEND_CRATES_ROOT", str(REPO_ROOT)))
+
+
+def _fixtures_cache_root() -> Path:
+    """Fixture extraction cache root (`~/.cache/dynamo/conformance-fixtures`
+    or `$XDG_CACHE_HOME/...`). `_common.sh` exports CONFORMANCE_FIXTURES_ROOT pointing
+    here; honor it first so staged renders and standalone runs agree."""
+    env = os.environ.get("CONFORMANCE_FIXTURES_ROOT")
+    if env:
+        return Path(env)
+    xdg = os.environ.get("XDG_CACHE_HOME")
+    base = Path(xdg) if xdg else Path.home() / ".cache"
+    return base / "dynamo/conformance-fixtures"
+
+
+_SRC_FIXTURES = _fixtures_cache_root() / "toolcalling/fixtures-batch-v1"
+# The resolver script stays in the repo (it's code, not a fixture).
+_RESOLVE_SRC_DIR = _FRONTEND_CRATES_ROOT / "conformance/utils/src"
+
+_VERSION_IMPLS = ("dynamo_v1", "vllm_python", "sglang_python")
+
+
+def _version_slug(version: str) -> str:
+    """CSS/DOM-safe token for a version, e.g. 0.5.12.post1 -> 0-5-12-post1."""
+    return re.sub(r"[^0-9A-Za-z]+", "-", version).strip("-")
+
+
+def _version_sort_key(version: str) -> tuple:
+    """Order versions like 0.5.12.post1 < 0.5.14 < 0.24.0 < 3.0.0."""
+    m = re.match(r"(\d+(?:\.\d+)*)(?:[.-]?post(\d+))?", version)
+    release = tuple(int(x) for x in m.group(1).split(".")) if m else ()
+    post = int(m.group(2)) if m and m.group(2) else 0
+    return (release, post)
+
+
+def _impl_versions() -> dict[str, list[str]]:
+    """Discover the versions present per impl from the fixture source dirs,
+    ascending. E.g. {"dynamo_v1": ["3.0.0"], "vllm_python": ["0.23.0", "0.24.0"], ...}."""
+    found: dict[str, list[str]] = {}
+    if not _SRC_FIXTURES.is_dir():
+        return found
+    for d in _SRC_FIXTURES.iterdir():
+        if not d.is_dir() or d.name == "inputs" or "-" not in d.name:
+            continue
+        impl, version = d.name.split("-", 1)
+        if impl in _VERSION_IMPLS:
+            found.setdefault(impl, []).append(version)
+    for impl in found:
+        found[impl] = sorted(set(found[impl]), key=_version_sort_key)
+    return {impl: found[impl] for impl in _VERSION_IMPLS if impl in found}
+
+
+def _pinned_versions(impl_versions: dict[str, list[str]]) -> dict[str, str]:
+    """Latest (pinned) version per impl = the default the radios select."""
+    return {impl: vers[-1] for impl, vers in impl_versions.items() if vers}
 
 
 def _build_family_inheritance(
