--- conformance/utils/src/tables/toolcalling/table.py
@@ -1,1278 +1,0 @@
-#!/usr/bin/env python3
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Generate the parity table (matrix of cell markers) from the YAML fixtures.
-
-================================================================================
-EXAMPLE OUTPUT (truncated; illustrative, NOT a snapshot of current fixtures
-— run the script for the real table):
-
-    | model          | parser     | 1 | 2.a | 2.b | 2.c | ... | 9 | 10 |
-    |---|---|:-:|:-:|:-:|:-:|:-:|:-:|:-:|
-    | **Top-N models** |   |   |   |   |   |   |   |   |
-    | Kimi K2.6      | kimi_k2    | = | =   | =   | VS  | ... | = | =  |
-    | gpt-oss        | harmony †  | S | S   | n/a | S?  | ... | = | S  |
-    | **Others** |   |   |   |   |   |   |   |   |
-    | Mistral series | mistral    | S | S   | n/a | VS  | ... | = | S  |
-
-================================================================================
-
-Reads every `tests/parity/toolcalling/fixtures/<family>/TOOLCALLING.batch*.yaml` and emits
-the table referenced in `tests/parity/README.md`.
-
-Cell markers (per peer, vllm + sglang):
-  =     peer block is `*d_<case>` anchor ref to dynamo (matches)
-  V/S   peer is a concrete inline block AND has `explanation:` (intentional)
-  V?/S? peer is a concrete inline block AND has no `explanation:` yet
-        (research-needed; we observed it but haven't classified it)
-  V✗/S✗ peer has `error: <substring>` (Python parser raised)
-  VS, V?S, VS✗, etc. — combinations
-  ·     Dynamo-only fixture; both peer blocks are `unavailable`
-  n/a   family/case doesn't apply
-  —     no fixture entry exists for this family/case yet
-
-Footnote markers `†` (no vLLM peer) and `§` (no SGLang peer) are auto-derived
-from `expected.<impl>.unavailable` across each family's cases.
-
-This module is a library: the conformance generator imports it for the batch/stream
-tool-calling table model. Render the page with `conformance/utils/render_table_v2.sh`.
-"""
-
-from __future__ import annotations
-
-import copy
-import html as html_lib
-import os
-import re
-import subprocess
-import sys
-import tempfile
-from pathlib import Path
-
-import yaml
-from jinja2 import Environment, FileSystemLoader, StrictUndefined
-
-import markers  # single source for comparison/leak/facts semantics (DIS-2477)
-from .. import common
-from ..common import TOP_N_TOOL_CALLING_FAMILIES as TOP_N_FAMILIES
-
-REPO_ROOT = Path(__file__).resolve().parents[2]
-FIXTURES = REPO_ROOT / "tests/parity/toolcalling/fixtures"
-TOOLCALLING_CASES_MD = REPO_ROOT / "lib/parsers/TOOLCALLING_CASES.md"
-TEMPLATE_DIR = REPO_ROOT / "tests/parity"
-
-# The versioned fixture source (inputs/ + per-impl <impl>-<version>/ dirs) lives in
-# the fixture extraction cache (from the in-repo LFS store). `_common.sh` exports
-# CONFORMANCE_FIXTURES_ROOT (the cache root); fall back to the standard cache path for
-# standalone runs. Used to power the per-impl version radios: we resolve each version
-# snapshot and re-run the load path so cell keys align exactly with the rendered
-# (pinned) table.
-_FRONTEND_CRATES_ROOT = Path(os.environ.get("FRONTEND_CRATES_ROOT", str(REPO_ROOT)))
-
-
-def _fixtures_cache_root() -> Path:
-    """Fixture extraction cache root (`~/.cache/dynamo/conformance-fixtures`
-    or `$XDG_CACHE_HOME/...`). `_common.sh` exports CONFORMANCE_FIXTURES_ROOT pointing
-    here; honor it first so staged renders and standalone runs agree."""
-    env = os.environ.get("CONFORMANCE_FIXTURES_ROOT")
-    if env:
-        return Path(env)
-    xdg = os.environ.get("XDG_CACHE_HOME")
-    base = Path(xdg) if xdg else Path.home() / ".cache"
-    return base / "dynamo/conformance-fixtures"
-
-
-_SRC_FIXTURES = _fixtures_cache_root() / "toolcalling/fixtures-batch-v1"
-# The resolver script stays in the repo (it's code, not a fixture).
-_RESOLVE_SRC_DIR = _FRONTEND_CRATES_ROOT / "conformance/utils/src"
-
-RUST_TOOL_CALLING_DIR = REPO_ROOT / "lib/parsers/src/tool_calling"
-
-# Row-label / visibility overrides keyed by tool calling family; ‡ is explained
-# by the legend note in the rendered page.
-_TOOL_CALLING_LABEL_OVERRIDES = {
-    "qwen3_coder": "Qwen 3 Coder / Nemotron V3‡",
-}
-# nemotron_nano: an alias for qwen3_coder, hide to avoid duplicate row
-# nemotron_deci: for older v2 nemotron models, hide to avoid confusion with nemotron v3 models
-_HIDDEN_TOOL_CALLING_FAMILIES = {"nemotron_deci", "nemotron_nano"}
-
-
-def _model_label_html(model: str) -> str:
-    """Escape a model label, styling any ‡ marker like the †/§ suffixes."""
-    return html_lib.escape(model).replace("‡", '<span class="parser-suffix">‡</span>')
-
-
-def _make_jinja_env() -> Environment:
-    env = Environment(
-        loader=FileSystemLoader(TEMPLATE_DIR),
-        trim_blocks=False,
-        lstrip_blocks=True,
-        undefined=StrictUndefined,
-    )
-    # Shared assets are template globals so every render site (parity table,
-    # reasoning parity, family-filtered) inlines them without repeating the kwargs.
-    env.globals["conformance_css"] = _read_asset("conformance.css")
-    env.globals["conformance_js"] = _read_asset("conformance.js")
-    # DIS-2434 JS view: builds the DOM from the model blob before conformance.js wires it.
-    env.globals["conformance_view_js"] = _read_asset("conformance_view.js")
-    return env
-
-
-def _read_asset(name: str) -> str:
-    """Inline a shared static asset (conformance.css / conformance.js) into the page.
-
-    The conformance table renders as a single self-contained HTML file that
-    inlines the `tests/parity/assets/` CSS+JS. Keeping one source avoids the
-    compare-bar/coloring logic drifting (it used to be duplicated inline)."""
-    return (TEMPLATE_DIR / "assets" / name).read_text(encoding="utf-8")
-
-
-def _commit_sha() -> str | None:
-    """HEAD SHA at table-generation time, or None if not in a git tree."""
-    try:
-        out = (
-            subprocess.check_output(
-                ["git", "-C", str(REPO_ROOT), "rev-parse", "HEAD"],
-                stderr=subprocess.DEVNULL,
-            )
-            .decode()
-            .strip()
-        )
-        return out or None
-    except (subprocess.CalledProcessError, FileNotFoundError):
-        return None
-
-
-# --- per-impl version snapshots (version radios) --------------------------------
-# The impls the version radios cover, in display order.
-_VERSION_IMPLS = ("dynamo_v1", "vllm_python", "sglang_python")
-
-
-def _version_slug(version: str) -> str:
-    """CSS/DOM-safe token for a version, e.g. 0.5.12.post1 -> 0-5-12-post1."""
-    return re.sub(r"[^0-9A-Za-z]+", "-", version).strip("-")
-
-
-def _version_sort_key(version: str) -> tuple:
-    """Order versions like 0.5.12.post1 < 0.5.14 < 0.24.0 < 3.0.0."""
-    m = re.match(r"(\d+(?:\.\d+)*)(?:[.-]?post(\d+))?", version)
-    release = tuple(int(x) for x in m.group(1).split(".")) if m else ()
-    post = int(m.group(2)) if m and m.group(2) else 0
-    return (release, post)
-
-
-def _impl_versions() -> dict[str, list[str]]:
-    """Discover the versions present per impl from the fixture source dirs,
-    ascending. E.g. {"dynamo_v1": ["3.0.0"], "vllm_python": ["0.23.0", "0.24.0"], ...}."""
-    found: dict[str, list[str]] = {}
-    if not _SRC_FIXTURES.is_dir():
-        return found
-    for d in _SRC_FIXTURES.iterdir():
-        if not d.is_dir() or d.name == "inputs" or "-" not in d.name:
-            continue
-        impl, version = d.name.split("-", 1)
-        if impl in _VERSION_IMPLS:
-            found.setdefault(impl, []).append(version)
-    for impl in found:
-        found[impl] = sorted(set(found[impl]), key=_version_sort_key)
-    return {impl: found[impl] for impl in _VERSION_IMPLS if impl in found}
-
-
-def _pinned_versions(impl_versions: dict[str, list[str]]) -> dict[str, str]:
-    """Latest (pinned) version per impl = the default the radios select."""
-    return {impl: vers[-1] for impl, vers in impl_versions.items() if vers}
-
-
-def _v1_peer_versions() -> dict[str, list[str]]:
-    """ALL captured peer versions (ascending) so both the older engines
-    (vLLM 0.23.0 / SGLang 0.5.12.post1) and the current ones (0.24.0 / 0.5.14)
-    are present and selectable in the compare bar. The oldest peer is the default
-    Compare candidate and newer ones default to the Others bucket — see
-    _candidate_items."""
-    return _impl_versions()
-
-
-def _version_status_map(mode: str) -> dict[tuple[str, str], dict[str, dict[str, str]]]:
-    """{(family, sub): {impl: {version_slug: overview_status}}}.
-
-    For each impl/version we resolve that version (other impls pinned) into a temp
-    flat tree and run the same `load_all_cases` path, so keys match the rendered
-    table exactly (including split-parent normalization). Status uses the same
-    `_overview_status` classifier as the pinned cells."""
-    impl_versions = _v1_peer_versions()
-    if not impl_versions:
-        return {}
-    resolver = _RESOLVE_SRC_DIR / "resolve_fixtures.py"
-    if not resolver.exists():
-        return {}
-    pinned = _pinned_versions(impl_versions)
-
-    global FIXTURES
-    result: dict[tuple[str, str], dict[str, dict[str, str]]] = {}
-    for impl, versions in impl_versions.items():
-        for version in versions:
-            slug = _version_slug(version)
-            # Resolve this impl@version with every other impl pinned, so the cell's
-            # color reflects only this impl's version change.
-            select = [
-                f"{other}-{version if other == impl else pinned[other]}"
-                for other in impl_versions
-            ]
-            with tempfile.TemporaryDirectory() as tmp:
-                subprocess.run(
-                    [sys.executable, str(resolver),
-                     "--fixtures-root", str(_SRC_FIXTURES),
-                     "--out", tmp, "--select", *select],
-                    check=True, capture_output=True,
-                )
-                saved = FIXTURES
-                FIXTURES = Path(tmp)
-                try:
-                    cases, _labels = load_all_cases(mode)
-                finally:
-                    FIXTURES = saved
-            for key, case in cases.items():
-                block = (case.get("expected") or {}).get(impl)
-                # Only `block` + `version` are consumed downstream (candidate charts
-                # + cmp payload in _v1_cell_model); the old per-slug status/marker
-                # glyphs were dead output — the JS view derives display from facts.
-                result.setdefault(key, {}).setdefault(impl, {})[slug] = {
-                    "block": block,
-                    "version": version,
-                }
-    return result
-
-
-def _build_family_inheritance(
-    refs: dict[str, tuple[str, int]],
-) -> dict[str, dict]:
-    """Derive each family's parser-inheritance map from config.rs + parsers.rs.
-
-    Detects:
-      • `ParserConfig::<Variant>(...)` — top-level backend variant
-      • `JsonParserType::<Sub>`         — Json sub-dispatch (Basic / DeepseekV3 / DeepseekV31)
-      • `Self::<factory>(...)`          — private factories (e.g. `deepseek_dsml`)
-      • `map.insert("alias", ToolCallConfig::<family>())` — aliases (parsers.rs)
-
-    Backend file is derived from the resolved (variant, sub_variant) tuple.
-    Returns `{family: {variant, sub_variant, factory, backend_file,
-    base_label, shared_with, aliases, filed_under_xml_misleading}}`.
-    """
-    cfg = (RUST_TOOL_CALLING_DIR / "config.rs").read_text()
-    pars_path = RUST_TOOL_CALLING_DIR / "parsers.rs"
-    pars = pars_path.read_text() if pars_path.exists() else ""
-
-    # Extract all ctor bodies (pub fn + fn) — captures private factories too.
-    ctor_pat = re.compile(
-        r"^\s*(?:pub )?fn (\w+)\([^)]*\)\s*->\s*Self\s*\{", re.MULTILINE
-    )
-    bodies: dict[str, str] = {}
-    for m in ctor_pat.finditer(cfg):
-        start = m.end()
-        depth, i = 1, start
-        while i < len(cfg) and depth > 0:
-            c = cfg[i]
-            if c == "{":
-                depth += 1
-            elif c == "}":
-                depth -= 1
-            i += 1
-        bodies[m.group(1)] = cfg[start : i - 1]
-
-    def _classify(body: str) -> tuple[str | None, str | None, str | None]:
-        vm = re.search(r"ParserConfig::(\w+)\b", body)
-        variant = vm.group(1) if vm else None
-        sub = None
-        if variant == "Json":
-            sm = re.search(r"JsonParserType::(\w+)\b", body)
-            sub = sm.group(1) if sm else "Basic"
-        fm = re.search(r"Self::(\w+)\(([^)]*)\)", body)
-        factory = f"{fm.group(1)}({fm.group(2).strip()})" if fm else None
-        return variant, sub, factory
-
-    backend_file = {
-        ("Json", "Basic"): "json/base_json_parser.rs",
-        ("Json", "DeepseekV3"): "json/deepseek_v3_parser.rs",
-        ("Json", "DeepseekV31"): "json/deepseek_v3_1_parser.rs",
-        ("Xml", None): "xml/parser.rs",
-        ("Pythonic", None): "pythonic/pythonic_parser.rs",
-        ("Harmony", None): "harmony/harmony_parser.rs",
-        ("Dsml", None): "dsml/parser.rs",
-        ("Glm47", None): "xml/glm47_parser.rs",
-        ("KimiK2", None): "xml/kimi_k2_parser.rs",
-        ("MiniMaxM3", None): "xml/minimax_m3_parser.rs",
-        ("Gemma4", None): "gemma4/parser.rs",
-    }
-    base_label = {
-        ("Json", "Basic"): "base_json_parser (JsonParserType::Basic)",
-        ("Json", "DeepseekV3"): "deepseek_v3_parser (JsonParserType::DeepseekV3)",
-        ("Json", "DeepseekV31"): "deepseek_v3_1_parser (JsonParserType::DeepseekV31)",
-        ("Xml", None): "xml::parser (shared XML base)",
-        ("Pythonic", None): "pythonic::parser (standalone)",
-        (
-            "Harmony",
-            None,
-        ): "harmony::parser (standalone; partial reuse of base_json's try_repair_truncated_json)",
-        ("Dsml", None): "dsml::parser (shared via deepseek_dsml() factory)",
-        ("Glm47", None): "glm47_parser (standalone; filed under xml/)",
-        ("KimiK2", None): "kimi_k2_parser (standalone; filed under xml/)",
-        ("MiniMaxM3", None): "minimax_m3_parser (standalone; filed under xml/)",
-        ("Gemma4", None): "gemma4::parser (standalone)",
-    }
-
-    out: dict[str, dict] = {}
-    for family in refs:
-        body = bodies.get(family)
-        if body is None:
-            continue
-        variant, sub, factory = _classify(body)
-        if variant is None and factory:
-            # Resolve through factory (e.g. deepseek_dsml)
-            fbody = bodies.get(factory.split("(")[0])
-            if fbody:
-                variant, sub, _ = _classify(fbody)
-        key = (variant, sub) if variant == "Json" else (variant, None)
-        out[family] = {
-            "variant": variant,
-            "sub_variant": sub,
-            "factory": factory,
-            "backend_file": backend_file.get(key, "unknown"),
-            "base_label": base_label.get(key, f"{variant}"),
-            "key": key,
-            "aliases": [],
-            "shared_with": [],
-            "filed_under_xml_misleading": False,
-        }
-
-    # Aliases from parsers.rs (only ones where alias name != family name).
-    alias_pat = re.compile(r'map\.insert\("([^"]+)",\s*ToolCallConfig::(\w+)\(\)\)')
-    alias_to_target: dict[str, str] = {}
-    for m in alias_pat.finditer(pars):
-        alias, fam = m.group(1), m.group(2)
-        if alias != fam and fam in out:
-            out[fam]["aliases"].append(alias)
-            alias_to_target[alias] = fam
-
-    # shared_with — other families with the same (variant, sub_variant).
-    by_key: dict[tuple, list[str]] = {}
-    for fam, info in out.items():
-        by_key.setdefault(info["key"], []).append(fam)
-    for fam, info in out.items():
-        info["shared_with"] = [s for s in by_key[info["key"]] if s != fam]
-        info["filed_under_xml_misleading"] = (
-            info["backend_file"].startswith("xml/") and info["variant"] != "Xml"
-        )
-
-    # Synthesize entries for alias-only families (e.g. nemotron_nano, qwen25).
-    # These are in `refs` (registered in parsers.rs) but have no ctor of their
-    # own — the alias `map.insert("nemotron_nano", ToolCallConfig::qwen3_coder())`
-    # routes to the target's config. The alias gets the target's full
-    # inheritance tree, plus `alias_of` so the tooltip can mark itself as a
-    # leaf under the target rather than as the target itself.
-    for alias, target in alias_to_target.items():
-        if alias in out or target not in out:
-            continue
-        tgt = out[target]
-        out[alias] = {
-            **tgt,
-            "alias_of": target,
-        }
-
-    return out
-
-
-def _build_family_to_rust_ref() -> dict[str, tuple[str, int]]:
-    """Scan the Rust source for each family's anchor point.
-
-    Two patterns:
-      `config.rs` :  `pub fn <family>() -> Self`               (parser config ctor)
-      `parsers.rs`:  `map.insert("<family>", ToolCallConfig::...);`  (aliases)
-
-    Config-ctor wins when the same family appears in both (the ctor is
-    the canonical definition; the registration is just plumbing). Aliases
-    (e.g. `nemotron_nano`, `qwen25`) only appear in `parsers.rs`.
-    Returns `{family: (filename, line)}`; line is 1-indexed.
-    """
-    out: dict[str, tuple[str, int]] = {}
-
-    config_rs = RUST_TOOL_CALLING_DIR / "config.rs"
-    if config_rs.exists():
-        pat = re.compile(r"^\s*pub fn (\w+)\(\)\s*->\s*Self\b")
-        for lineno, line in enumerate(config_rs.read_text().splitlines(), 1):
-            m = pat.match(line)
-            if m:
-                out[m.group(1)] = ("config.rs", lineno)
-
-    parsers_rs = RUST_TOOL_CALLING_DIR / "parsers.rs"
-    if parsers_rs.exists():
-        pat = re.compile(r'^\s*map\.insert\("([^"]+)",\s*ToolCallConfig::')
-        for lineno, line in enumerate(parsers_rs.read_text().splitlines(), 1):
-            m = pat.match(line)
-            if m and m.group(1) not in out:
-                out[m.group(1)] = ("parsers.rs", lineno)
-
-    return out
-
-
-BATCH_SUB_CASE_GROUPS = [
-    ("Single-call", ("1.a", "1.b", "1.c", "1.d")),
-    ("Core", ("1", "3", "9", "9.a", "9.b")),
-    ("Multi-call", ("2.a", "2.b", "2.c", "2.d", "2.e", "10")),
-    (
-        "Malformed / recovery",
-        (
-            "4.a",
-            "4.b",
-            "4.c",
-            "4.d",
-            "4.e",
-            "4.f",
-            "5.a",
-            "5.b",
-            "5.c",
-            "5.d",
-            "5.e",
-            "5.f",
-            "5.g",
-            # 5.h is streaming-only (no batch.5.h) but the v2 stream tab reuses the
-            # batch taxonomy, so it must be ordered here beside 5.g — otherwise it
-            # sorts to the far right as an "unknown" case, away from its 5.* siblings.
-            "5.h",
-        ),
-    ),
-    (
-        "Args",
-        (
-            "6.a",
-            "6.b",
-            "6.c",
-            "7.a",
-            "7.b",
-            "7.c",
-            "7.d",
-            "7.e",
-            "7.f",
-        ),
-    ),
-    ("Text interleaving", ("8.a", "8.b", "8.c", "8.d")),
-    ("Unknown tools", ("13", "13.a", "13.c")),
-    (
-        "String contents",
-        ("30", "30.a", "30.b", "30.c", "31", "31.a", "31.b"),
-    ),
-]
-
-SPLIT_PARENT_SUBCASES = {
-    # Once a taxonomy bucket has leaf cases, the matrix should render only the
-    # leaves. Existing parent fixtures still carry useful expectations/reasons
-    # for parser families that have not been rewritten to leaf IDs yet.
-    "1": ("1.a",),
-    "9": ("9.a",),
-    "30": ("30.a", "30.b", "30.c"),
-    "31": ("31.a", "31.b"),
-    "13": ("13.a",),
-}
-
-STREAM_SUB_CASE_GROUPS = [
-    ("Single-call", ("1.a", "1.b")),
-    ("Multi-call", ("2",)),
-    ("Partial-token", ("3",)),
-    ("Termination", ("4.a", "4.b", "4.c")),
-]
-
-SUB_CASE_GROUPS_BY_MODE = {
-    "batch": BATCH_SUB_CASE_GROUPS,
-    "stream": STREAM_SUB_CASE_GROUPS,
-}
-
-_SUB_CASE_GROUP_KEY_BY_LABEL_BY_MODE = {
-    mode: {
-        label: re.sub(r"[^a-z0-9]+", "_", label.lower()).strip("_")
-        for label, _subs in groups
-    }
-    for mode, groups in SUB_CASE_GROUPS_BY_MODE.items()
-}
-
-_SUB_CASE_GROUP_KEY_BY_SUB_BY_MODE = {
-    mode: {
-        sub: _SUB_CASE_GROUP_KEY_BY_LABEL_BY_MODE[mode][label]
-        for label, subs in groups
-        for sub in subs
-    }
-    for mode, groups in SUB_CASE_GROUPS_BY_MODE.items()
-}
-
-
-def _display_order(mode: str) -> dict[str, tuple[int, int]]:
-    return {
-        sub: (group_idx, sub_idx)
-        for group_idx, (_label, subs) in enumerate(SUB_CASE_GROUPS_BY_MODE[mode])
-        for sub_idx, sub in enumerate(subs)
-    }
-
-
-def _group_index_by_sub(mode: str) -> dict[str, int]:
-    return {
-        sub: group_idx
-        for group_idx, (_label, subs) in enumerate(SUB_CASE_GROUPS_BY_MODE[mode])
-        for sub in subs
-    }
-
-
-def _group_by_sub(mode: str) -> dict[str, str]:
-    return {sub: label for label, subs in SUB_CASE_GROUPS_BY_MODE[mode] for sub in subs}
-
-
-def _natural_sub_sort_key(sub: str) -> tuple[int, str]:
-    """`8.a` → (8, 'a'); `9` → (9, '')."""
-    parts = sub.split(".")
-    return (int(parts[0]), parts[1] if len(parts) > 1 else "")
-
-
-def _sub_sort_key(mode: str, sub: str) -> tuple[int, int, int, str]:
-    """Sort known cases by semantic display group, future cases naturally last."""
-    display_order = _display_order(mode).get(sub)
-    if display_order is not None:
-        group_idx, sub_idx = display_order
-        return (0, group_idx, sub_idx, "")
-    num, suffix = _natural_sub_sort_key(sub)
-    return (1, num, 0, suffix)
-
-
-def _subcase_band_class(mode: str, sub: str) -> str:
-    group_idx = _group_index_by_sub(mode).get(sub, len(SUB_CASE_GROUPS_BY_MODE[mode]))
-    return f"case-band-{group_idx % 2}"
-
-
-def _subcase_group_key(mode: str, sub: str) -> str:
-    return _SUB_CASE_GROUP_KEY_BY_SUB_BY_MODE[mode].get(sub, "other")
-
-
-def _discover_sub_cases(mode: str, cases: dict) -> list[str]:
-    """Union of sub-case IDs across all loaded fixtures, in stable order."""
-    return sorted(
-        {sub for _fam, sub in cases.keys()}, key=lambda s: _sub_sort_key(mode, s)
-    )
-
-
-def _normalize_split_parent_cases(cases: dict) -> dict:
-    """Render split taxonomy buckets as leaf cases only.
-
-    Some older fixture files still define parent buckets such as
-    `TOOLCALLING.batch.30`, while newer files define leaf buckets such as
-    `TOOLCALLING.batch.30.a`. For display, parent+leaf duplication is confusing:
-    once any leaf exists for a parent bucket, the table should show only the
-    leaf columns. Parent entries are copied into missing leaf cells so their
-    existing expectations or n/a reasons remain visible until the YAML itself
-    is migrated.
-    """
-    all_subs = {sub for _fam, sub in cases.keys()}
-    active_split_parents = {
-        parent
-        for parent, children in SPLIT_PARENT_SUBCASES.items()
-        if any(child in all_subs for child in children)
-    }
-    if not active_split_parents:
-        return cases
-
-    normalized = dict(cases)
-    families = {fam for fam, _sub in cases.keys()}
-    for family in families:
-        for parent in active_split_parents:
-            parent_key = (family, parent)
-            parent_case = normalized.get(parent_key)
-            if parent_case is None:
-                continue
-            for child in SPLIT_PARENT_SUBCASES[parent]:
-                child_key = (family, child)
-                if child_key in normalized:
-                    continue
-                child_case = copy.deepcopy(parent_case)
-                child_case["__case_id"] = f"TOOLCALLING.batch.{child}"
-                child_case["__synthetic_from_case_id"] = parent_case.get("__case_id")
-                normalized[child_key] = child_case
-            del normalized[parent_key]
-    return normalized
-
-
-def _derive_no_peer_sets(cases: dict) -> tuple[set[str], set[str]]:
-    """Families where every case marks the engine `unavailable`.
-
-    Used to render the † (no vLLM peer) and § (no SGLang peer) footnote
-    markers next to a family's name. A family qualifies when every case
-    in every fixture file under that family has
-    `expected.<impl>.unavailable: <reason>` recorded — i.e. the wrapper
-    rejected the family for that parser in `capture_toolcalling_outputs.py`.
-    """
-    by_family: dict[str, list[dict]] = {}
-    for (fam, _sub), case in cases.items():
-        by_family.setdefault(fam, []).append(case)
-
-    def all_unavail(fam_cases: list[dict], impl: str) -> bool:
-        expected_cases = [c for c in fam_cases if isinstance(c.get("expected"), dict)]
-        if not expected_cases:
-            return False
-        for c in expected_cases:
-            block = c.get("expected", {}).get(impl)
-            if not isinstance(block, dict) or "unavailable" not in block:
-                return False
-        return True
-
-    no_vllm = {fam for fam, cs in by_family.items() if all_unavail(cs, "vllm_python")}
-    no_sglang = {fam for fam, cs in by_family.items() if all_unavail(cs, "sglang_python")}
-    return no_vllm, no_sglang
-
-
-def family_suffix(fam: str, no_vllm: set[str], no_sglang: set[str]) -> str:
-    suff = ""
-    if fam in no_vllm:
-        suff += "†"
-    if fam in no_sglang:
-        suff += "§"
-    return suff
-
-
-def load_all_cases(mode: str) -> tuple[dict[tuple[str, str], dict], dict[str, str]]:
-    """Load every fixture YAML for one parser mode.
-
-    Returns `(cases, labels)`:
-      cases  — `{(family, sub_case_id): case_data}`; each case dict gets
-               `__fixture_path` (relative to this script) and `__case_id`
-               annotations for the HTML renderer.
-      labels — `{family: model_label}` collected from the fixtures' doc-level
-               `model_label:` field. Falls back to the family ID if a fixture
-               doesn't declare one.
-    """
-    cases: dict[tuple[str, str], dict] = {}
-    labels: dict[str, str] = {}
-    script_dir = Path(__file__).resolve().parent
-    for fp in sorted(FIXTURES.glob(f"*/TOOLCALLING.{mode}*.yaml")):
-        doc = yaml.safe_load(fp.read_text())
-        if doc.get("mode") != mode:
-            continue
-        family = doc["family"]
-        rel = common.LINKS["toolcalling_fixtures"] + str(fp.relative_to(FIXTURES))
-        if "model_label" in doc:
-            labels.setdefault(family, doc["model_label"])
-        for cid, case in doc["cases"].items():
-            case["__family"] = family
-            sub = cid.replace(f"TOOLCALLING.{mode}.", "")
-            case["__fixture_path"] = rel
-            case["__case_id"] = cid
-            cases[(family, sub)] = case
-    return _normalize_split_parent_cases(cases), labels
-
-
-def _build_display_groups(
-    cases: dict, labels: dict[str, str]
-) -> tuple[list[tuple[str, str]], list[tuple[str, str]]]:
-    """Return `(top_n, others)` as `[(label, family), ...]` lists.
-
-    Top-N: families listed in `TOP_N_FAMILIES`, in that exact order.
-    Others: every YAML-discovered family not in TOP_N, sorted by label.
-    Missing labels fall back to the family ID.
-    """
-    families = {
-        fam for fam, _ in cases.keys() if fam not in _HIDDEN_TOOL_CALLING_FAMILIES
-    }
-
-    def label_of(fam: str) -> str:
-        return _TOOL_CALLING_LABEL_OVERRIDES.get(fam, labels.get(fam, fam))
-
-    top_n = [(label_of(f), f) for f in TOP_N_FAMILIES if f in families]
-    other_fams = sorted(
-        families - set(TOP_N_FAMILIES), key=lambda f: label_of(f).lower()
-    )
-    others = [(label_of(f), f) for f in other_fams]
-    return top_n, others
-
-
-# DIS-2477: comparison / leak / status semantics are single-sourced in markers.py.
-# This table used to fork its own copies of these here (some diverged: no
-# parser-error branch, a hardcoded leak regex, no JSON-arg canonicalization). Route
-# the local names to the shared implementations so there is one source of truth.
-peer_status = markers.peer_status
-_explanation = markers._explanation
-_dynamo_tool_call_leak = markers._dynamo_tool_call_leak
-_block_tool_call_leaks = markers._block_tool_call_leaks
-_overview_status = markers._overview_status
-
-
-def cell_for(case: dict | None) -> str:
-    if case is None:
-        return "—"
-    dyn = case.get("expected", {}).get("dynamo_v1")
-    if not isinstance(dyn, dict):
-        return "n/a"
-    v_kind, v_unknown = peer_status(case, dyn, "vllm_python")
-    s_kind, s_unknown = peer_status(case, dyn, "sglang_python")
-
-    parts: list[str] = []
-    if v_kind == "div":
-        parts.append("V?" if v_unknown else "V")
-    elif v_kind == "err":
-        parts.append("V✗")
-    if s_kind == "div":
-        parts.append("S?" if s_unknown else "S")
-    elif s_kind == "err":
-        parts.append("S✗")
-
-    # `explanation:` on the `expected.dynamo` block flags Dynamo's own output as
-    # leaking tool call markup only when Dynamo also leaves residual
-    # `normal_text`. Dynamo can have non-leak reasons for dropped malformed
-    # markup, so don't mark those as `↯`.
-    if isinstance(dyn, dict) and _dynamo_tool_call_leak(dyn):
-        if v_kind == "unavail" and s_kind == "unavail":
-            return "↯·"
-        if parts:
-            return "↯" + "".join(parts)
-        return "↯"
-
-    if parts:
-        return "".join(parts)
-    if v_kind == "unavail" and s_kind == "unavail":
-        return "·"
-    return "="
-
-
-_LEGEND_MD = (
-    "**Legend:** "
-    "`=` all captured peers match Dynamo · "
-    "`·` Dynamo-only fixture (both peers unavailable) · "
-    "`V`/`S` divergence (V = vLLM, S = SGLang; intentional, has `explanation:`) · "
-    "`?` research-needed suffix (e.g. V?, S? — diverges with no `explanation:` yet) · "
-    "`↯` Dynamo leaks tool call markup into `normal_text` "
-    "(`expected.dynamo.reason:` carries the explanation) · "
-    "`✗` parser exception (e.g. V✗, S✗ — Python parser raised) · "
-    "`n/a` not applicable · "
-    "`—` missing fixture coverage · "
-    "`†` (tool calling parser column) = no vLLM peer parser for this family · "
-    "`§` (tool calling parser column) = no SGLang peer parser for this family."
-    "\n\n"
-    "`‡` Nemotron V3 (Ultra) reuses the qwen3_coder tool calling parser; "
-    "Nemotron V1 / V2 (DeciLM) is removed from the chart for being an older "
-    "generation, but the nemotron_deci parser is still supported."
-)
-
-
-_IMPL_DISPLAY = {"dynamo_v1": "Dynamo", "vllm_python": "vLLM", "sglang_python": "SGLang"}
-
-
-def _parser_inheritance_tooltip_html(
-    family: str,
-    info: dict,
-    ctor_ref: tuple[str, int] | None,
-    no_vllm: set[str] | None = None,
-    no_sglang: set[str] | None = None,
-) -> str:
-    """Rich `.ttip` tooltip for the tool calling parser column.
-
-    Keep this field-list shape aligned with the reasoning parser column tooltip
-    so both tables explain "effective parser/backend -> row family" the same
-    way. `ctor_ref` is unused here (was for older field-based layout) — kept
-    for API stability with `_parser_cell_html`.
-    """
-    del ctor_ref
-
-    variant = info["variant"] or "?"
-    sub_variant = info["sub_variant"]
-    backend_file = info["backend_file"]
-    factory = info["factory"]
-    alias_of = info.get("alias_of")  # set when this family is an alias-only entry
-
-    head_parts = [f"ParserConfig::{variant}"]
-    if sub_variant:
-        head_parts[-1] = f"ParserConfig::{variant}::{sub_variant}"
-    bf_href = html_lib.escape(f"{common.LINKS['toolcalling_src']}{backend_file}")
-    bf_link = f'<a href="{bf_href}">{html_lib.escape(backend_file)}</a>'
-
-    anchor = alias_of or family
-    shared_family = sorted([anchor] + info["shared_with"])
-    effective_backend = _shared_backend_short(info) or family
-
-    implementation = f"{html_lib.escape(head_parts[0])} -> {bf_link}"
-    if factory:
-        factory_name = factory.split("(", 1)[0]
-        implementation += html_lib.escape(f" (factory: {factory_name})")
-
-    tooltip_lines = [
-        "Tool calling parser family from fixture YAML.",
-        f"Tool calling parser row: {html_lib.escape(family)}",
-        f"Effective parser/backend: {html_lib.escape(effective_backend)}",
-        f"Dynamo implementation: {implementation}",
-    ]
-    if info["shared_with"]:
-        tooltip_lines.append(
-            "Shared implementation family: " + html_lib.escape(", ".join(shared_family))
-        )
-    if alias_of:
-        tooltip_lines.append(f"Alias of: {html_lib.escape(alias_of)}")
-    if info["aliases"]:
-        tooltip_lines.append(
-            "Registered aliases: " + html_lib.escape(", ".join(info["aliases"]))
-        )
-
-    peer_notes: list[str] = []
-    if no_vllm and family in no_vllm:
-        peer_notes.append("no vLLM peer parser")
-    if no_sglang and family in no_sglang:
-        peer_notes.append("no SGLang peer parser")
-    if peer_notes:
-        tooltip_lines.append("Peer availability: " + ", ".join(peer_notes))
-
-    if info["filed_under_xml_misleading"]:
-        tooltip_lines.append(
-            "Note: filed under xml/ but does not use the shared xml::parser; "
-            f"it has its own ParserConfig::{html_lib.escape(variant)} variant."
-        )
-    tooltip_lines.extend(_tool_parser_tree_lines(family, info, effective_backend))
-
-    if effective_backend == family:
-        head_text = f"`{family}`"
-    else:
-        head_text = f"`{effective_backend}` (row: `{family}`)"
-    return (
-        '<div class="ttip">'
-        f'<div class="ttip-head">{html_lib.escape(head_text)}</div>'
-        f'<pre class="ttip-pre">{"".join(line + chr(10) for line in tooltip_lines).rstrip()}</pre>'
-        "</div>"
-    )
-
-
-_SHARED_BACKEND_SHORT = {
-    ("Json", "Basic"): "base_json",
-    ("Xml", None): "xml",
-    ("Dsml", None): "dsml",
-}
-
-
-def _tool_parser_tree_lines(
-    family: str,
-    info: dict,
-    effective_backend: str,
-) -> list[str]:
-    alias_of = info.get("alias_of")
-    anchor = alias_of or family
-    aliases = info["aliases"]
-    if not info["shared_with"] and not aliases and effective_backend == family:
-        return []
-
-    fam_list = sorted([anchor] + info["shared_with"])
-    lines = ["", "Shared implementation tree:"]
-    root_label = html_lib.escape(effective_backend)
-    if effective_backend == family:
-        root_label = f"<strong>{root_label}</strong>"
-    lines.append(f"{root_label} (effective parser/backend)")
-
-    for i, fam in enumerate(fam_list):
-        is_last_fam = i == len(fam_list) - 1
-        branch = "└── " if is_last_fam else "├── "
-        fam_label = html_lib.escape(fam)
-        if fam == family and not alias_of:
-            fam_label = f"<strong>{fam_label}</strong>"
-        lines.append(f"{branch}{fam_label}")
-
-        if fam == anchor and aliases:
-            cont = "    " if is_last_fam else "│   "
-            for j, alias in enumerate(aliases):
-                alast = j == len(aliases) - 1
-                ab = "└── " if alast else "├── "
-                alias_label = html_lib.escape(alias)
-                if alias_of and alias == family:
-                    alias_label = f"<strong>{alias_label}</strong>"
-                lines.append(f"{cont}{ab}{alias_label} (alias)")
-
-    return lines
-
-
-def _shared_backend_short(info: dict | None) -> str | None:
-    if info and info["shared_with"]:
-        return _SHARED_BACKEND_SHORT.get(info["key"])
-    return None
-
-
-def _parser_cell_html(
-    family: str,
-    refs: dict[str, tuple[str, int]],
-    no_vllm: set[str],
-    no_sglang: set[str],
-    inheritance: dict[str, dict],
-) -> str:
-    suff = family_suffix(family, no_vllm, no_sglang)
-    row_label = html_lib.escape(family)
-    if suff:
-        row_label += f'<span class="parser-suffix">{html_lib.escape(suff)}</span>'
-    ref = refs.get(family)
-    info = inheritance.get(family)
-    ttip = (
-        _parser_inheritance_tooltip_html(family, info, ref, no_vllm, no_sglang)
-        if info
-        else ""
-    )
-
-    # Shared-backend rows should read as implementation -> fixture family,
-    # e.g. `xml -> minimax_m2` and `xml -> qwen3_coder`. Standalone parsers
-    # keep the public family name as the primary label.
-    short = _shared_backend_short(info)
-    if short:
-        label = html_lib.escape(short)
-        base_suffix = f'<span class="parser-base">→ {row_label}</span>'
-    else:
-        label = row_label
-        base_suffix = ""
-
-    # Family-name link points to the **actual parser code** (backend_file from
-    # the inheritance map), not to the config-ctor location in config.rs. The
-    # ctor location is still referenced in the inheritance tooltip body when
-    # useful (factory calls). For families with no inheritance info, fall back
-    # to the refs entry (config.rs or parsers.rs).
-    if info and info["backend_file"] != "unknown":
-        href = f"{common.LINKS['toolcalling_src']}{info['backend_file']}"
-    elif ref is not None:
-        href = f"{common.LINKS['toolcalling_src']}{ref[0]}"
-    else:
-        return (
-            f'<td class="parser" data-col-hide-group="parser">'
-            f"{label}{base_suffix}{ttip}</td>"
-        )
-    return (
-        f'<td class="parser" data-col-hide-group="parser">'
-        f'<a href="{href}">{label}</a>{base_suffix}{ttip}</td>'
-    )
-
-
-def _parse_subcase_descriptions(mode: str) -> dict[str, str]:
-    """Parse `lib/parsers/TOOLCALLING_CASES.md` for per-case descriptions.
-
-    The Quick-reference section has one-liner bullets for top-level cases
-    (`TOOLCALLING.<mode>.1` …); the deeper per-case sections
-    contain multi-line bullets for sub-cases (`2.a`, `4.c`, etc.). Both
-    look like `- **`TOOLCALLING.<mode>.X`** <desc>`, where the bullet body may
-    wrap across indented continuation lines. Returns
-    `{"1": "...", "2.a": "...", ...}`.
-    """
-    if not TOOLCALLING_CASES_MD.exists():
-        return {}
-    pat = re.compile(
-        rf"\*\*`TOOLCALLING\.{re.escape(mode)}" rf"\.([0-9]+(?:\.[a-z])?)`\*\*\s+(.+)"
-    )
-    out: dict[str, str] = {}
-    lines = TOOLCALLING_CASES_MD.read_text(encoding="utf-8").splitlines()
-    i = 0
-    while i < len(lines):
-        m = pat.search(lines[i])
-        if not m:
-            i += 1
-            continue
-        sub = m.group(1)
-        body_parts = [m.group(2).strip()]
-        # Join indented continuation lines until blank / next bullet / unindented.
-        j = i + 1
-        while j < len(lines):
-            nxt = lines[j]
-            if not nxt.strip():
-                break
-            if not nxt.startswith(" "):
-                break
-            if pat.search(nxt):
-                break
-            body_parts.append(nxt.strip())
-            j += 1
-        desc = " ".join(body_parts).rstrip(".")
-        out.setdefault(sub, desc)
-        i = j
-    return out
-
-
-def _subcase_group_label(mode: str, sub: str) -> str:
-    return _group_by_sub(mode).get(sub, "Other")
-
-
-def _subcase_runs(mode: str, sub_cases: list[str]) -> list[list[str]]:
-    runs: list[list[str]] = []
-    start = 0
-    while start < len(sub_cases):
-        label = _subcase_group_label(mode, sub_cases[start])
-        end = start + 1
-        while (
-            end < len(sub_cases) and _subcase_group_label(mode, sub_cases[end]) == label
-        ):
-            end += 1
-        runs.append(sub_cases[start:end])
-        start = end
-    return runs
-
-
-def _glossary_groups(
-    mode: str, descriptions: dict[str, str], sub_cases: list[str]
-) -> list[dict[str, object]]:
-    if not descriptions:
-        return []
-    return [
-        {
-            "label": _subcase_group_label(mode, run[0]),
-            "rows": [
-                (
-                    sub,
-                    descriptions.get(sub) or descriptions.get(sub.split(".")[0]) or "",
-                )
-                for sub in run
-            ],
-        }
-        for run in _subcase_runs(mode, sub_cases)
-    ]
-
-
-_IMPL_RADIO_LABEL = {"dynamo_v1": "Dynamo", "vllm_python": "vLLM", "sglang_python": "SGLang"}
-
-
-# Candidate label base per parity impl: "<Engine> <Runtime>". Parity runs the v1
-# parsers; the standardized label is "<base> <version> (<mode>)" (e.g.
-# "Dynamo Rust 3.0.0 (batch)"), matching the conformance page. Dynamo's parser is a
-# Rust crate (dynamo-parsers 3.0.0).
-_PARITY_CAND_BASE = {"dynamo_v1": "Dynamo Rust", "vllm_python": "vLLM Python", "sglang_python": "SGLang Python"}
-
-
-def _candidate_items(mode: str = "batch") -> list[dict[str, str]]:
-    """Ordered comparison candidates for the compare model: Dynamo, then vLLM/SGLang
-    versions ascending. Each: {key, impl, version, slug, short, label, default_bucket}.
-    Labels are "<Engine> <Runtime> <version> (<mode>)", e.g. "Dynamo Rust 3.0.0
-    (batch)" / "vLLM Python 0.24.0 (stream)".
-
-    Default layout: A (reference) = Dynamo; B (compare with) = each peer's OLDEST
-    (v1-era) version — this is the legacy page, so the older engines are the default
-    comparison; C (others) = each peer's newer versions, present but not shown until
-    dragged into Compare."""
-    impl_versions = _v1_peer_versions()
-    oldest = {impl: (vers[0] if vers else None) for impl, vers in impl_versions.items()}
-    out: list[dict[str, str]] = []
-    first = True
-    for impl in _VERSION_IMPLS:
-        base = _PARITY_CAND_BASE.get(impl, _IMPL_RADIO_LABEL.get(impl, impl))
-        # Parity runs the v1 Dynamo crate (dynamo-parsers 3.x), so Dynamo reads
-        # "Dynamo v1 Rust 3.0.0 (batch)" / "(stream)"; peers have no crate split.
-        if impl == "dynamo_v1":
-            eng, _, rt = base.partition(" ")  # -> "Dynamo v1 Rust"
-            base = f"{eng} v1 {rt}".strip()
-        for v in impl_versions.get(impl, []):
-            slug = _version_slug(v)
-            if first:
-                bucket = "A"
-                first = False
-            elif v == oldest.get(impl):
-                bucket = "B"
-            else:
-                bucket = "C"
-            out.append(
-                {
-                    "key": f"{impl}-{slug}",
-                    "impl": impl,
-                    "version": v,
-                    "slug": slug,
-                    "short": base,
-                    "label": f"{base} {v} ({mode})",
-                    "default_bucket": bucket,
-                }
-            )
-    return out
-
-
-def _candidate_cmp_json(case: dict | None) -> dict | None:
-    """Versioned per-cell compare payload {`<impl>-<slug>`: {sig, leak, na}}, built by
-    the shared markers.cmp_model (canonical JSON-arg signatures single-sourced there —
-    the v1 fork used to byte-compare serialized args and flag false divergences)."""
-    ver = (case or {}).get("__ver_status") if isinstance(case, dict) else None
-    if not ver:
-        return None
-    blocks = {
-        f"{impl}-{slug}": info.get("block")
-        for impl, by_slug in ver.items()
-        for slug, info in by_slug.items()
-    }
-    return markers.cmp_model(blocks) if blocks else None
-
-
-def _compute_stats(
-    cases: dict, sub_cases: list[str], families: list[str]
-) -> dict[str, int]:
-    """Aggregate cell outcomes across the (family × sub_case) grid."""
-    s = {
-        "families": len(families),
-        "sub_cases": len(sub_cases),
-        "slots": len(families) * len(sub_cases),
-        "real": 0,
-        "parity": 0,
-        "dynamo_only": 0,
-        "documented": 0,
-        "research": 0,
-        "errors": 0,
-        "na": 0,
-        "missing": 0,
-    }
-    for fam in families:
-        for sub in sub_cases:
-            case = cases.get((fam, sub))
-            text = cell_for(case)
-            if text == "—":
-                s["missing"] += 1
-                continue
-            if text == "n/a":
-                s["na"] += 1
-                continue
-            s["real"] += 1
-            if text == "=":
-                s["parity"] += 1
-            elif text in {"D", "·"}:
-                s["dynamo_only"] += 1
-            elif "!" in text or "✗" in text:
-                s["errors"] += 1
-            elif "↯" in text:
-                s["documented"] += 1
-            elif "?" in text:
-                s["research"] += 1
-            else:
-                s["documented"] += 1
-    return s
-
-
-# ===== Structured JSON model builder ===========================================
-# Same-schema model tab as the v2 toolcalling path (model.make_cell). Comparison
-# semantics (status / cmp / facts) are single-sourced in markers.py.
-import model  # noqa: E402  (schema + cell normalizer; leaf module staged alongside)
-
-
-def _cand_engine_group_tc(key: str) -> str:
-    for prefix in ("dynamo", "vllm", "sglang"):
-        if key.startswith(prefix):
-            return prefix
-    return key
-
-
-def _v1_output_block_model(blk: object) -> dict | None:
-    if not isinstance(blk, dict):
-        return None
-    out: dict[str, Any] = {}
-    if "unavailable" in blk:
-        out["unavailable"] = blk["unavailable"]
-    if "error" in blk:
-        out["error"] = blk["error"]
-    if "calls" in blk or "normal_text" in blk:
-        out["calls"] = blk.get("calls") or []
-        out["normal_text"] = blk.get("normal_text") or ""
-    expl = _explanation(blk)
-    if expl:
-        out["explanation"] = expl
-    return out or None
-
-
-def _v1_columns_model(mode: str, sub_cases: list[str], descriptions: dict[str, str]) -> tuple[list[dict], list[dict]]:
-    groups: list[dict] = []
-    cols: list[dict] = []
-    for run in _subcase_runs(mode, sub_cases):
-        gk = _subcase_group_key(mode, run[0])
-        groups.append({"key": gk, "label": _subcase_group_label(mode, run[0]),
-                       "band": _subcase_band_class(mode, run[0]), "span": len(run)})
-        for sub in run:
-            cols.append({"sub": sub, "group_key": gk, "band": _subcase_band_class(mode, sub),
-                         "label": sub,
-                         "desc": descriptions.get(sub) or descriptions.get(sub.split(".")[0]) or ""})
-    return groups, cols
-
-
-def _v1_cell_model(case: dict | None, mode: str, family: str, sub: str,
-                   cand_label_by_key: dict[str, str]) -> dict:
-    group_key = _subcase_group_key(mode, sub)
-    band = _subcase_band_class(mode, sub)
-    if case is None:
-        return model.missing_cell(sub, family, group_key, band,
-                                  head=f"TOOLCALLING.{mode}.{sub}")
-    ver = case.get("__ver_status") or {}
-    cmp = _candidate_cmp_json(case)
-    candidates = []
-    for impl in ("dynamo_v1", "vllm_python", "sglang_python"):
-        for slug, info in (ver.get(impl) or {}).items():
-            key = f"{impl}-{slug}"
-            candidates.append({
-                "key": key, "label": cand_label_by_key.get(key, key),
-                "impl": _cand_engine_group_tc(key), "version": info.get("version"),
-                "parse_mode": mode, "block": _v1_output_block_model(info.get("block")),
-                "leak": isinstance(info.get("block"), dict) and _block_tool_call_leaks(info["block"]),
-            })
-    # Structured comparison facts, single-sourced in markers (was a stub here with only
-    # `status` filled; markers computes real present/agrees/intentional/reason/leak).
-    facts = markers.comparison_facts(case, markers.BATCH_IMPL_KEYS, "dynamo_v1")
-    fp = case.get("__fixture_path", "")
-    href = common.fixture_href(fp) if fp else None
-    tooltip = {
-        "head": f"{case.get('__case_id','')} — {family}",
-        "description": case.get("description") or "",
-        "input": {"kind": "text" if case.get("model_text") else None,
-                  "text": case.get("model_text"), "chunks": None, "family": family},
-        "candidates": candidates, "baseline": None, "reasons": [],
-        "dynamo_notes": [], "refs": [], "leak_note": None, "na_note": None,
-    }
-    return model.make_cell(kind="cell", case_id=case.get("__case_id"), family=family,
-                           sub=sub, col_group=group_key, band=band, fixture_href=href,
-                           status=_overview_status(case, "dynamo_v1"), cmp=cmp,
-                           facts=facts, tooltip=tooltip)
-
-
-def build_model_panel(mode: str, active: bool = False) -> dict:
-    """v1 toolcalling tab as a structured model dict (schema shared with v2/reasoning)."""
-    cases, labels = load_all_cases(mode)
-    ver_status = _version_status_map(mode)
-    for key, case in cases.items():
-        if isinstance(case, dict) and key in ver_status:
-            case["__ver_status"] = ver_status[key]
-    sub_cases = _discover_sub_cases(mode, cases)
-    no_vllm, no_sglang = _derive_no_peer_sets(cases)
-    top_n, others = _build_display_groups(cases, labels)
-    descriptions = _parse_subcase_descriptions(mode)
-    refs = _build_family_to_rust_ref()
-    inheritance = _build_family_inheritance(refs)
-    column_groups, cols = _v1_columns_model(mode, sub_cases, descriptions)
-    cand_items = _candidate_items(mode)
-    cand_label_by_key = {c["key"]: c["label"] for c in cand_items}
-
-    def row_model(model_label: str, family: str) -> dict:
-        cells = {
-            sub: _v1_cell_model(cases.get((family, sub)), mode, family, sub, cand_label_by_key)
-            for sub in sub_cases
-        }
-        return {
-            "section": None,
-            "model_label": model_label,
-            "model_label_html": _model_label_html(model_label),
-            "family": family,
-            "parser": {"html": _parser_cell_html(family, refs, no_vllm, no_sglang, inheritance)},
-            "cells": cells,
-        }
-
-    rows: list[dict] = []
-    if top_n:
-        rows.append({"section": "Top-N models", "model_label": "Top-N models",
-                     "model_label_html": "", "family": None, "parser": None, "cells": {}})
-        rows.extend(row_model(m, f) for m, f in top_n)
-    if others:
-        rows.append({"section": "Others", "model_label": "Others",
-                     "model_label_html": "", "family": None, "parser": None, "cells": {}})
-        rows.extend(row_model(m, f) for m, f in others)
-
-    all_families = [f for _, f in top_n] + [f for _, f in others]
-    return {
-        "id": f"tab-toolcalling-{mode}",
-        "kind": "toolcalling",
-        "mode": mode,
-        "column_groups": column_groups,
-        "columns": cols,
-        "rows": rows,
-        "stats": _compute_stats(cases, sub_cases, all_families),
-        "glossary": _glossary_groups(mode, descriptions, sub_cases),
-        "candidates": cand_items,
-    }
-
-
-if __name__ == "__main__":
-    main()
